requirements.txt
lancedb>=0.29
sentence-transformers>=5.0
pyyaml>=6.0
